+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - Suggest awaiting promise before using it when types mismatch. https://github.com/rescript-lang/rescript/pull/7498
 - Complete from `RegExp` stdlib module for regexes. https://github.com/rescript-lang/rescript/pull/7425
+- Allow oneliner formatting when including module with single type alias. https://github.com/rescript-lang/rescript/pull/7502
 
 # 12.0.0-alpha.13